requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
aiohttp==3.9.1
pandas==2.1.3
pandas-ta==0.3.14b0
python-dotenv==1.0.0